# Batch-input to egg