… src/core/patch.rkt
… src/core/preprocess.rkt
… src/core/programs.rkt
… src/core/simplify.rkt
… src/core/taylor.rkt