This PR introduces new changes to `egg-herbie` pipeline.
Particularly speaking, `batch`-type input has been implemented for `egg`.
The reason behind this change is: 
Previously, `egg` received a `listof Expr` as an input for `egraph`. These expressions were converted to `batch` structure using internal `egraph` functions (`egraph_add_enode`).
We have noticed that this conversion from `listof Expr` to `batch-egraph` is redundant as we are planning to switch to `batch` structure for expressions inside Herbie.

Instead of doing what is described above, we convert this `listof Expr` to a `batch` on Herbie side and pass nodes to `egraph`.

The new strategy brings some speedup at `egraph-add-exprs` (370s -> 258s) and some slowdown on conversions from `progs->batch` (159s -> 210s). In addition, some extra conversions between `batch` and `listof Expr` are introduced at `simplify-batch` (~45s). Overall, the runtime is expected to be the same (and it is).

Tech details of PR:
The PR changes structure of `egraph-run-schedule`. Instead of storing `exprs`, it stores `batch` and corresponding `roots` that are to be inserted. 
The new structure of `egraph-run-schedule` matches the new `insert to egg` strategy.