@AYadrov 2024-09-09
Some differences in the output accuracy from nightly has been observed ([`main`](https://nightly.cs.washington.edu/reports/herbie/1725877903:nightly:main:02e7ec5a37/) vs [`artem-batch-egg-input`](https://nightly.cs.washington.edu/reports/herbie/1725913682:nightly:artem-batch-egg-input:a97bd55497/)).

After debugging, the reason behind the differences is believed to be a change of the order of nodes for `egraph_add_node` function. Previously the order was recursion-based (starting from a root of expression), while now the order is the same as the order of nodes inside `batch`.